requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-wtf>=1.1.1
flask-caching>=2.1.0
flask-compress>=1.14
flask-limiter>=3.5.0
werkzeug>=3.0.0
pypdf2>=3.0.0
reportlab>=4.0.0
pillow>=10.0.0
asgiref>=3.9.1
uvicorn[standard]>=0.35.0
cryptography>=41.0.0
pycryptodome>=3.19.0
bcrypt>=4.0.1
email-validator>=1.3.1
gunicorn>=21.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
aiofiles>=23.0.0
aiocache>=0.12.0
alembic>=1.13.0
psutil>=5.9.0
ldap3>=2.9.0
